move seedhack into common
